feat: refactors and type annotations p3 (queue ui)
